demo2/requirements.txt
# Demo2 Package Requirements
# Core dependencies for AI analysis functionality

# OpenAI Agents and related packages
openai-agents==0.3.3
openai==1.109.1

# Supabase integration
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0

# MCP (Model Context Protocol)
mcp==1.17.0

# Data processing
pandas==2.3.3
numpy==2.3.4

# Environment and configuration
python-dotenv==1.1.1

# HTTP and networking
httpx==0.28.1
requests==2.32.5

# Utilities
rich==14.2.0
typer==0.19.2
